[client] 仕様書項目Form Table分離
